Refac

+++ pacman-ui-fx-3d/src/main/java/de/amr/games/pacman/ui/fx/v3d/entity/WallData.java
@@ -9,8 +9,8 @@ package de.amr.games.pacman.ui.fx.v3d.entity;
  */
 public class WallData {
     byte type; // see FloorPlan
-    int x;
-    int y;
-    int numBricksX;
-    int numBricksY;
+    short x;
+    short y;
+    short numBricksX;
+    short numBricksY;
 }
